UVE: configure-once example kernels + so.b.ndc dimension-index fix
Three new whole-program examples that show the point of UVE — the loop nest lives in stream descriptors configured once, with no scalar address arithmetic or per-iteration reconfiguration:

- GemmTests: dense GEMM via 3D streams with stride-0 repeat dimensions; the entire i/j/k nest is five instructions (mac + ndc branch + store + nc branch).
- TriangularSumModifierTests: lower-triangular sum as a single 2D stream with an ss.app.mod {Size, Inc} modifier — the two-instruction contrast to the per-row-reconfiguring LowerTriangularSumTests.
- SparseDotProductTests: sparse·dense dot product gathering x[col[k]] through an ss.sta.ld.w_inds IndSource stream driving an ss.app.ind {Offset, Set} modifier — first ISA-level exercise of indirect stream modifiers.

Writing GEMM exposed a so.b.ndc bug: Spike encodes funct3 = D-1 counting dimensions from the OUTERMOST (EODTable.at(funct3), dimensions[0] = outermost), but the pipeline passed funct3 straight into the engine's innermost-first DimDone index — the branch checked the wrong dimension and the innermost dim was unreachable (funct3=0 collides with so.b.nc). OooeTrain now remaps via StreamingEngine.DimensionCount when syncing DimDone; the flag stays keyed by raw funct3. No prior integration test used so.b.ndc, so this was never caught.

Also recorded in TODO.md: Spike builds descriptors outermost-first (ss.end = innermost) while Horologium builds innermost-first — a program-level config-order divergence to align in a future pass.

+++ Tests/Orrery/StreamingEngineTests.cs
@@ -507,7 +507,7 @@ public class StreamingEngineTests {
         //   total: 9 elements
 
         const int nWords = 16;
-        var mem = new FlatMemory((int)(0x200 + nWords * 4));
+        var mem = new FlatMemory(0x200 + nWords * 4);
         for (var i = 0; i < nWords; i++) mem.Load((ulong)(i * 4), BitConverter.GetBytes((uint)(i + 1)));
         mem.Load(0x100, BitConverter.GetBytes(3u));
         mem.Load(0x104, BitConverter.GetBytes(2u));
@@ -561,7 +561,7 @@ public class StreamingEngineTests {
         // total: 4 + 7 + 7 = 18 elements
 
         const int nWords = 32;
-        var mem = new FlatMemory((int)(0x300 + nWords * 4));
+        var mem = new FlatMemory(0x300 + nWords * 4);
         for (var i = 0; i < nWords; i++) mem.Load((ulong)(i * 4), BitConverter.GetBytes((uint)(i + 1)));
         mem.Load(0x200, BitConverter.GetBytes(2u));
         mem.Load(0x204, BitConverter.GetBytes(3u));
